app/requirements/all.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
click==8.3.1
fastapi==0.124.4
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
librt==0.7.4
Mako==1.3.10
MarkupSafe==3.0.3
mypy==1.19.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pydantic==2.12.5
pydantic_core==2.41.5
pyhumps==3.8.0
python-dotenv==1.2.1
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
